1378: Renamed classes variables to analyticArithmetic and eventCount
