+++ packages/indicators/src/Builder/AnalyticArithmeticBuilder/AnalyticArithmeticBuilder.ts
@@ -13,7 +13,7 @@ import { Builder } from '../Builder';
 import { createBuilder } from '../createBuilder';
 import { fetchAnalytics, validateConfig, evaluateFormulaToNumber } from '../helpers';
 import {
-  ArithmeticConfig,
+  AnalyticArithmeticConfig,
   configValidators,
   DefaultValue,
   getAggregationListByCode,
@@ -30,7 +30,7 @@ type BuilderConfig = {
   readonly defaultValues: Record<string, DefaultValue>;
 };
 
-const indicatorToBuilderConfig = (indicatorConfig: ArithmeticConfig): BuilderConfig => {
+const indicatorToBuilderConfig = (indicatorConfig: AnalyticArithmeticConfig): BuilderConfig => {
   const { defaultValues = {}, parameters = [], ...otherFields } = indicatorConfig;
 
   return {
@@ -41,7 +41,7 @@ const indicatorToBuilderConfig = (indicatorConfig: ArithmeticConfig): BuilderCon
   };
 };
 
-export class ArithmeticBuilder extends Builder {
+export class AnalyticArithmeticBuilder extends Builder {
   private configCache: BuilderConfig | null = null;
 
   private paramBuildersByCodeCache: Record<string, Builder> | null = null;
@@ -69,7 +69,7 @@ export class ArithmeticBuilder extends Builder {
 
   validateConfig = () => {
     const { config } = this.indicator;
-    validateConfig<ArithmeticConfig>(config, configValidators);
+    validateConfig<AnalyticArithmeticConfig>(config, configValidators);
     return config;
   };
