+++ packages/indicators/src/Builder/AnalyticArithmeticBuilder/config/validate.ts
