+++ packages/indicators/src/Builder/AnalyticArithmeticBuilder/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ */
+
+export { AnalyticArithmeticBuilder } from './AnalyticArithmeticBuilder';
+export { AnalyticArithmeticConfig } from './config';
