--- packages/indicators/src/Builder/ArithmeticBuilder/index.ts
@@ -1,7 +1,0 @@
-/**
- * Tupaia
- * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
- */
-
-export { ArithmeticBuilder } from './ArithmeticBuilder';
-export { ArithmeticConfig } from './config';
