--- packages/indicators/src/Builder/CountEventBuilder/index.ts
@@ -1,7 +1,0 @@
-/**
- * Tupaia
- * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
- */
-
-export { CountEventBuilder } from './CountEventBuilder';
-export { CountEventConfig } from './config';
