+++ packages/indicators/src/Builder/EventCountBuilder/config/types.ts
@@ -12,7 +12,7 @@ export type DefaultValue = number | string | 'undefined';
 /**
  * A flexible config schema that can be provided as an input to the indicator
  */
-export type CountEventConfig = {
+export type EventCountConfig = {
   readonly formula: string;
   readonly programCode: string;
   readonly defaultValues?: Record<string, DefaultValue>;
