+++ packages/indicators/src/Builder/EventCountBuilder/config/validate.ts
