+++ packages/indicators/src/Builder/EventCountBuilder/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ */
+
+export { EventCountBuilder } from './EventCountBuilder';
+export { EventCountConfig } from './config';
