+++ packages/indicators/src/Builder/builders.ts
@@ -3,10 +3,10 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 
-import { ArithmeticBuilder } from './ArithmeticBuilder';
-import { CountEventBuilder } from './CountEventBuilder';
+import { AnalyticArithmeticBuilder } from './AnalyticArithmeticBuilder';
+import { EventCountBuilder } from './EventCountBuilder';
 
 export const builders = {
-  arithmeticAnalytic: ArithmeticBuilder,
-  countEvent: CountEventBuilder
+  analyticArithmetic: AnalyticArithmeticBuilder,
+  eventCount: EventCountBuilder
 };
