+++ packages/indicators/src/__tests__/Builder/AnalyticArithmeticBuilder/AnalyticArithmeticBuilder.fixtures.ts
