+++ packages/indicators/src/__tests__/Builder/AnalyticArithmeticBuilder/AnalyticArithmeticBuilder.test.ts
