+++ packages/indicators/src/__tests__/__integration/fixtures/helpers.ts
@@ -52,5 +52,5 @@ export const entriesToArithmeticIndicators = (entries: IndicatorEntries) =>
     if (config.parameters) {
       newConfig.parameters = entriesToArithmeticIndicators(config.parameters as IndicatorEntries);
     }
-    return { code, builder: 'arithmetic', config: newConfig };
+    return { code, builder: 'analyticArithmetic', config: newConfig };
   });
